Fix #178
