+++ .travis.yml
@@ -1,3 +1,3 @@
 language: dart
-script: dart tool/grind.dart buildbot
+script: pub run grinder:grinder
 sudo: false
