+++ pubspec.yaml
@@ -21,5 +21,4 @@ dev_dependencies:
   unittest: any
 
 executables:
-  grind:
-  grinder:
+  grind: grinder
